Add react-router-dom & create pages
