+++ client/src/pages/home/Home.css
@@ -1,10 +1,3 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
 .logo {
   height: 5.5em;
   padding: 0.5em;
@@ -35,10 +28,6 @@
   a:nth-of-type(2) .logo {
     animation: logo-spin infinite 20s linear;
   }
-}
-
-.read-the-docs {
-  color: #888;
 }
 
 .weight-track-0 {
